requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.2
pyarrow==17.0.0
PyYAML==6.0.2
playwright==1.47.0